ghosts will not kill pacman in debug mode

+++ Pacman/GameObjects/Sprites/Ghosts/Ghost.cs
@@ -380,7 +380,7 @@ namespace FormsPixelGameEngine.GameObjects.Sprites.Ghosts
                 if (mode == Mode.FRIGHTENED)
                     Eat();
                  
-                else if (mode != Mode.EATEN)
+                else if (mode != Mode.EATEN && !game.Debug)
                     pacman.Kill();
             }
